refactor: DefaultController 구현
