+++ tomcat/src/main/java/nextstep/jwp/controller/DefaultController.java
@@ -1,5 +1,6 @@
-package org.apache.catalina.handler;
+package nextstep.jwp.controller;
 
+import org.apache.catalina.handler.AbstractController;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
@@ -7,11 +8,12 @@ public class DefaultController extends AbstractController {
 
     @Override
     protected HttpResponse doGet(final HttpRequest request) {
-        return null;
+        return new HttpResponse.Builder(request).ok()
+            .messageBody(getStaticResource(request.getUrl())).build();
     }
 
     @Override
     protected HttpResponse doPost(final HttpRequest request) {
-        throw new IllegalArgumentException("Unsupported Method for this path: \"/\"");
+        throw new IllegalArgumentException("Unsupported Method for this path: " + request.getPath());
     }
 }
