+++ tomcat/src/main/java/org/apache/catalina/handler/AbstractController.java
@@ -1,5 +1,11 @@
 package org.apache.catalina.handler;
 
+import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.nio.file.Files;
+import java.util.Objects;
+
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
@@ -13,11 +19,22 @@ public abstract class AbstractController implements Controller {
         return doPost(request);
     }
 
+    protected abstract HttpResponse doGet(final HttpRequest request);
+
+    protected abstract HttpResponse doPost(final HttpRequest request);
+
     protected boolean isGetRequest(final HttpRequest request) {
         return request.isGetRequest();
     }
 
-    protected abstract HttpResponse doGet(final HttpRequest request);
+    protected String getStaticResource(final URL url) {
+        try {
+            return Files.readString(new File(Objects.requireNonNull(url)
+                .getFile())
+                .toPath());
+        } catch (IOException e) {
+            throw new IllegalArgumentException("No such resource");
+        }
+    }
 
-    protected abstract HttpResponse doPost(final HttpRequest request);
 }
